chore: eip1272 comaptible message signature validation
